backend/requirements.txt
python-dotenv>=1.0.1
pyyaml>=6.0.2

numpy>=1.26.0
scikit-learn>=1.4.0
rank-bm25>=0.2.2

openai>=1.40.0

streamlit>=1.36.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.8.0

faiss-cpu>=1.8.0

pdfplumber>=0.11.0
pypdf>=5.0.0